# swagger/api.yaml
openapi: 3.0.0
info:
  title: Social Media API
  description: |
    Social Media Backend API built with NestJS and MongoDB.
    This API provides authentication, user management, posts, comments, stories, groups, pages, and messaging functionality.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /posts:
    get:
      tags:
        - Posts
      summary: Get all posts
      description: Returns a list of all posts with pagination
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

    post:
      tags:
        - Posts
      summary: Create a new post
      description: Creates a new post
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '201':
          description: Post created successfully
        '401':
          description: Unauthorized
  /posts/user/{userId}:
    get:
      tags:
        - Posts
      summary: Get posts by user
      description: Returns all posts by a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of user posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get post by ID
      description: Returns a specific post by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found

    patch:
      tags:
        - Posts
      summary: Update post
      description: Updates a post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePost'
      responses:
        '200':
          description: Post updated successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Post not found

    delete:
      tags:
        - Posts
      summary: Delete post
      description: Deletes a post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Post not found
  /posts/{id}/like:
    post:
      tags:
        - Posts
      summary: Like a post
      description: Likes or unlikes a post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post liked/unliked successfully
        '401':
          description: Unauthorized
        '404':
          description: Post not found
  /posts/{id}/share:
    post:
      tags:
        - Posts
      summary: Share a post
      description: Shares a post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post shared successfully
        '401':
          description: Unauthorized
        '404':
          description: Post not found
  /posts/{id}/repost:
    post:
      tags:
        - Posts
      summary: Repost a post
      description: Creates a repost of an existing post
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption:
                  type: string
      responses:
        '201':
          description: Repost created successfully
        '401':
          description: Unauthorized
        '404':
          description: Post not found
  /posts/upload:
    post:
      tags:
        - Posts
      summary: Upload files for posts
      description: Uploads images or videos for posts
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Files uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        url:
                          type: string
                        originalName:
                          type: string
                        size:
                          type: integer
                        type:
                          type: string
                          enum: [image, video]
...
  /stories:
    get:
      tags:
        - Stories
      summary: Get all stories
      description: Returns a list of all stories with pagination
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Story'

    post:
      tags:
        - Stories
      summary: Create a new story
      description: Creates a new story
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStory'
      responses:
        '201':
          description: Story created successfully
        '401':
          description: Unauthorized
  /stories/user/{userId}:
    get:
      tags:
        - Stories
      summary: Get stories by user
      description: Returns all stories by a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of user stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Story'
  /stories/{id}:
    get:
      tags:
        - Stories
      summary: Get story by ID
      description: Returns a specific story by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Story found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '404':
          description: Story not found

    delete:
      tags:
        - Stories
      summary: Delete story
      description: Deletes a story
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Story deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Story not found
  /stories/{id}/view:
    post:
      tags:
        - Stories
      summary: View a story
      description: Marks a story as viewed by the current user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Story viewed successfully
        '401':
          description: Unauthorized
        '404':
          description: Story not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: string
        avatar:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Post:
      type: object
      properties:
        _id:
          type: string
        author:
          $ref: '#/components/schemas/User'
        content:
          type: string
        media:
          type: array
          items:
            type: string
        likes:
          type: array
          items:
            type: string
        shares:
          type: integer
        comments:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePost:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        media:
          type: array
          items:
            type: string
        groupId:
          type: string
        pageId:
          type: string
    UpdatePost:
      type: object
      properties:
        content:
          type: string
        media:
          type: array
          items:
            type: string
...
    Story:
      type: object
      properties:
        _id:
          type: string
        author:
          $ref: '#/components/schemas/User'
        media:
          type: string
        views:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateStory:
      type: object
      required:
        - media
      properties:
        media:
          type: string
        caption:
          type: string
...